requirements.txt
-e .
jupyterlab
notebook
jupyter
seaborn
matplotlib
pandas
numpy
pyyaml